refactor: 메뉴단으로 나누는 것보다 기능 관리 단에서 다시 리펙토링
